requirements.txt
pandas
openai
openpyxl
httpx
python-dotenv
